Partially revert "Bump Qt to 5.15 LTS"
This mostly reverts commit 252ad53377013b02a03b8b8a1a877be04181745e.

The part which bumps the AppVeyor build to use Qt 5.15 is fine but there
is no apparent reason to *require* the update. Building with older
versions seems fine, and we should only require new versions when we
have an actual reason to do so, which should be clearly stated in the
commit message.